requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
opencv-python==4.5.1.48
Pillow==8.0.1
pandas==1.2.1
streamlit==0.78.0
streamlit-webrtc==0.7.2
matplotlib==3.3.3
tqdm==4.56.0
PyYAML==5.3.1
seaborn==0.11.1